Refactors fundfile path handling
load and save now take paths to the file instead of the directory,
meaning that the file can be named anything the user wanted.
This change also eliminates two more instances of unnecessary cloning.
The paths are now constructed outside of these methods, so they don't
need to take ownership of the path values.